added tests, changed type of column

+++ lib/models/image.dart
@@ -13,7 +13,7 @@ class _Image {
   @Column(nullable: true)
   String resolution;
 
-  double version; // 0 for original, the rest customisable. 1.1, 1.2, 2.1 etc.
+  int version;
 
   @Relate(#versions)
   ImageMetaData imageMetaData;
